::before and ::after WIP 2
