+++ packages/blitz-dom/src/layout/construct.rs
@@ -12,7 +12,7 @@ use style::{
     },
     shared_lock::StylesheetGuards,
     values::{
-        computed::Display,
+        computed::{Content, ContentItem, Display},
         specified::box_::{DisplayInside, DisplayOutside},
     },
 };
@@ -33,6 +33,8 @@ pub(crate) fn collect_layout_children(
     layout_children: &mut Vec<usize>,
     anonymous_block_id: &mut Option<usize>,
 ) {
+    flush_pseudo_elements(doc, container_node_id);
+
     if let Some(el) = doc.nodes[container_node_id].raw_dom_data.downcast_element() {
         // Handle text inputs
         let tag_name = el.name.local.as_ref();
@@ -204,7 +206,121 @@ pub(crate) fn collect_layout_children(
         }
 
         _ => {
+            if let Some(before) = doc.nodes[container_node_id].before.clone() {
+                layout_children.push(before);
+            }
             layout_children.extend_from_slice(&doc.nodes[container_node_id].children);
+            if let Some(after) = doc.nodes[container_node_id].after.clone() {
+                layout_children.push(after);
+            }
+        }
+    }
+}
+
+fn flush_pseudo_elements(doc: &mut Document, node_id: usize) {
+    let (before_style, after_style, before_node_id, after_node_id) = {
+        let node = &doc.nodes[node_id];
+
+        let before_node_id = node.before.clone();
+        let after_node_id = node.after.clone();
+
+        // Note: yes these are kinda backwards
+        let style_data = node.stylo_element_data.borrow();
+        let before_style = style_data
+            .as_ref()
+            .and_then(|d| d.styles.pseudos.as_array()[1].clone());
+        let after_style = style_data
+            .as_ref()
+            .and_then(|d| d.styles.pseudos.as_array()[0].clone());
+
+        (before_style, after_style, before_node_id, after_node_id)
+    };
+
+    for (idx, pe_style, pe_node_id) in [
+        (1, before_style, before_node_id),
+        (0, after_style, after_node_id),
+    ] {
+        // Drop pseudo element node if style no longer exists
+        if pe_node_id.is_some() && pe_style.is_none() {
+            doc.remove_and_drop_node(pe_node_id.unwrap());
+            doc.nodes[node_id].set_pe_by_index(idx, None);
+            // doc.nodes[node_id]
+            //     .layout_children
+            //     .borrow_mut()
+            //     .as_mut()
+            //     .unwrap()
+            //     .retain(|id| *id != node_id);
+        }
+
+        // Create pseudo element node if it doesn't exist but should
+        if pe_node_id.is_none() && pe_style.is_some() {
+            const NAME: QualName = QualName {
+                prefix: None,
+                ns: ns!(html),
+                local: local_name!("div"),
+            };
+            let new_node_id = doc.create_node(NodeData::AnonymousBlock(ElementNodeData::new(
+                NAME,
+                Vec::new(),
+            )));
+
+            let content = &pe_style.as_ref().unwrap().get_counters().content;
+            if let Content::Items(item_data) = content {
+                let items = &item_data.items[0..item_data.alt_start];
+                match &items[0] {
+                    ContentItem::String(owned_str) => {
+                        // dbg!(&owned_str);
+                        let text_node_id = doc.create_text_node(&owned_str);
+                        doc.nodes[new_node_id].children.push(text_node_id);
+                        *doc.nodes[new_node_id].layout_children.borrow_mut() = Some(vec![text_node_id]);
+                        // let parley_style = stylo_to_parley::style(&*pe_style.as_ref().unwrap());
+
+                        // // Create a parley tree builder
+                        // let mut builder =
+                        //     doc.layout_ctx
+                        //         .tree_builder(&mut doc.font_ctx, doc.viewport.scale(), &parley_style);
+
+                        // // Set whitespace collapsing mode
+                        // let collapse_mode = stylo_to_parley::white_space_collapse(pe_style.as_ref().unwrap().clone_white_space_collapse());
+                        // builder.set_white_space_mode(collapse_mode);
+
+                        // builder.push_text(&owned_str);
+
+                        // let (layout, text) = builder.build();
+                        // let text_layout = Box::new(TextLayout { text, layout });
+
+                        // doc.nodes[new_node_id]
+                        //     .raw_dom_data
+                        //     .downcast_element_mut()
+                        //     .unwrap()
+                        //     .inline_layout_data = Some(text_layout);
+                    },
+                    ContentItem::Counter(_custom_ident, _t) => { /* todo */ },
+                    ContentItem::Counters(_custom_ident, _owned_str, _t) => { /* todo */ },
+                    ContentItem::OpenQuote => { /* todo */ },
+                    ContentItem::CloseQuote => { /* todo */ },
+                    ContentItem::NoOpenQuote => { /* todo */ },
+                    ContentItem::NoCloseQuote => { /* todo */ },
+                    ContentItem::Attr(_attr) => { /* todo */ },
+                    ContentItem::Image(_) => { /* todo */ },
+                }
+            }
+
+            doc.nodes[node_id].set_pe_by_index(idx, Some(new_node_id));
+            // doc.nodes[node_id]
+            //     .layout_children
+            //     .borrow_mut()
+            //     .as_mut()
+            //     .unwrap()
+            //     .push(new_node_id);
+        }
+
+        // Sync style to pseudo element
+        if let Some(pe_style) = pe_style {
+            let mut element_data = ElementData::default();
+            element_data.styles.primary = Some(pe_style);
+            element_data.set_restyled();
+            *doc.nodes[node_id].stylo_element_data.borrow_mut() = Some(element_data);
         }
     }
 }
@@ -399,10 +515,7 @@ fn collect_complex_layout_children(
     hide_whitespace: bool,
     needs_wrap: impl Fn(NodeKind, DisplayOutside) -> bool,
 ) {
-    // Take children array from node to avoid borrow checker issues.
-    let children = std::mem::take(&mut doc.nodes[container_node_id].children);
-
-    for child_id in children.iter().copied() {
+    doc.iter_children_and_pseudos_mut(container_node_id, |child_id, doc| {
         // Get node kind (text, element, comment, etc)
         let child_node_kind = doc.nodes[child_id].raw_dom_data.kind();
 
@@ -428,7 +541,7 @@ fn collect_complex_layout_children(
         //
         // Also hide all-whitespace flexbox children as these should be ignored
         if child_node_kind == NodeKind::Comment || (hide_whitespace && is_whitespace_node) {
-            continue;
+            return;
         }
         // Recurse into `Display::Contents` nodes
         else if display_inside == DisplayInside::Contents {
@@ -477,10 +590,7 @@ fn collect_complex_layout_children(
             *anonymous_block_id = None;
             layout_children.push(child_id);
         }
-    }
-
-    // Put children array back
-    doc.nodes[container_node_id].children = children;
+    });
 }
 
 fn create_text_editor(doc: &mut Document, input_element_id: usize, is_multiline: bool) {
@@ -559,7 +669,7 @@ pub(crate) fn build_inline_layout(
         .unwrap_or(WhiteSpaceCollapse::Collapse);
     builder.set_white_space_mode(collapse_mode);
 
-    //Render position-inside list items
+    // Render position-inside list items
     if let Some(ListItemLayout {
         marker,
         position: ListItemLayoutPosition::Inside,
@@ -573,11 +683,30 @@ pub(crate) fn build_inline_layout(
         }
     };
 
+    if let Some(before_id) = root_node.before.clone() {
+        dbg!(before_id);
+        build_inline_layout_recursive(
+            &mut builder,
+            &doc.nodes,
+            before_id,
+            collapse_mode,
+            root_line_height,
+        );
+    }
     for child_id in root_node.children.iter().copied() {
         build_inline_layout_recursive(
             &mut builder,
             &doc.nodes,
             child_id,
+            collapse_mode,
+            root_line_height,
+        );
+    }
+    if let Some(after_id) = root_node.after.clone() {
+        build_inline_layout_recursive(
+            &mut builder,
+            &doc.nodes,
+            after_id,
             collapse_mode,
             root_line_height,
         );
@@ -675,11 +804,29 @@ pub(crate) fn build_inline_layout(
 
                             builder.push_style_span(style);
 
+                            if let Some(before_id) = node.before.clone() {
+                                build_inline_layout_recursive(
+                                    builder,
+                                    nodes,
+                                    before_id,
+                                    collapse_mode,
+                                    root_line_height,
+                                );
+                            }
                             for child_id in node.children.iter().copied() {
                                 build_inline_layout_recursive(
                                     builder,
                                     nodes,
                                     child_id,
+                                    collapse_mode,
+                                    root_line_height,
+                                );
+                            }
+                            if let Some(after_id) = node.after.clone() {
+                                build_inline_layout_recursive(
+                                    builder,
+                                    nodes,
+                                    after_id,
                                     collapse_mode,
                                     root_line_height,
                                 );
